重构Models结构
